social-media-traffic/requirements.txt
psycopg2-binary
faker